clippy

+++ src/motifs/lcs.rs
@@ -9,7 +9,7 @@ pub(crate) fn lcs(fasta: &[FASTA]) -> String {
         if DNA.contains(&&*char::from_u32(i).unwrap().to_string()) {
             i += 1;
         }
-        tree.add_string(f.dna.clone(), char::from_u32(i as u32).unwrap());
+        tree.add_string(f.dna.clone(), char::from_u32(i).unwrap());
         i += 1
     });
     tree.longest_common_substring_all()
